pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "A travel planning application with MCP integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "httpx",
    "uvicorn",
    "mcp[cli]",
    "pydantic",
    "python-dotenv",
    "starlette",
    "streamlit",
    "langchain-groq",
    "langchain-core",
    "langchain-mcp-adapters",
    "apify-client>=1.9.3",
    "amadeus>=12.0.0",
    "google>=3.0.0",
    "maps>=5.1.1",
    "googlemaps>=4.10.0",
]

[project.scripts]
start = "src.server:run"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
